Refactor types into separate module
